+++ packages/beacon-state-transition/src/allForks/index.ts
@@ -1,6 +1,5 @@
 export * from "./signatureSets";
 export * from "./stateTransition";
-export * from "./util";
 export * from "./block";
 export * from "./epoch";
