--- packages/beacon-state-transition/src/allForks/util/epochStakeSummary.ts
@@ -1,5 +1,0 @@
-export interface IEpochStakeSummary {
-  sourceStakeByIncrement: number;
-  targetStakeByIncrement: number;
-  headStakeByIncrement: number;
-}
