--- packages/beacon-state-transition/src/allForks/util/index.ts
@@ -1,11 +1,0 @@
-export * from "./attestation";
-export * from "./attesterStatus";
-export * from "./cachedBeaconState";
-export * from "./epochContext";
-export * from "./epochProcess";
-export * from "./epochShuffling";
-export * from "./epochStakeSummary";
-export * from "./flat";
-export * from "./syncCommitteeCache";
-export * from "./pubkeyCache";
-export * from "./weakSubjectivity";
