+++ packages/beacon-state-transition/src/cache/balanceList.ts
