+++ packages/beacon-state-transition/src/cache/cachedInactivityScoreList.ts
