+++ packages/beacon-state-transition/src/cache/effectiveBalanceIncrements.ts
