+++ packages/beacon-state-transition/src/cache/epochProcess.ts
@@ -1,0 +1,110 @@
+import {Epoch, ValidatorIndex} from "@chainsafe/lodestar-types";
+import {IAttesterStatus} from "../util/attesterStatus";
+
+/**
+ * Pre-computed disposable data to process epoch transitions faster at the cost of more memory.
+ *
+ * The AttesterStatus (and FlatValidator under status.validator) objects and
+ * EpochStakeSummary are tracked in the IEpochProcess and made available as additional context in the
+ * epoch transition.
+ */
+export interface IEpochProcess {
+  prevEpoch: Epoch;
+  currentEpoch: Epoch;
+  /**
+   * This is sum of active validators' balance in eth.
+   */
+  totalActiveStakeByIncrement: number;
+  /** For altair */
+  baseRewardPerIncrement: number;
+  prevEpochUnslashedStake: {
+    sourceStakeByIncrement: number;
+    targetStakeByIncrement: number;
+    headStakeByIncrement: number;
+  };
+  currEpochUnslashedTargetStakeByIncrement: number;
+  /**
+   * Indices which will receive the slashing penalty
+   * ```
+   * v.withdrawableEpoch === currentEpoch + EPOCHS_PER_SLASHINGS_VECTOR / 2
+   * ```
+   * There's a practical limitation in number of possible validators slashed by epoch, which would share the same
+   * withdrawableEpoch. Note that after some count exitChurn would advance the withdrawableEpoch.
+   * ```
+   * maxSlashedPerSlot = SLOTS_PER_EPOCH * (MAX_PROPOSER_SLASHINGS + MAX_ATTESTER_SLASHINGS * bits)
+   * ```
+   * For current mainnet conditions (bits = 128) that's `maxSlashedPerSlot = 8704`.
+   * For less than 327680 validators, churnLimit = 4 (minimum possible)
+   * For exitChurn to overtake the slashing delay, there should be
+   * ```
+   * churnLimit * (EPOCHS_PER_SLASHINGS_VECTOR / 2 - 1 - MAX_SEED_LOOKAHEAD)
+   * ```
+   * For mainnet conditions that's 16364 validators. So the limiting factor is the max operations on the block. Note
+   * that on average indicesToSlash must contain churnLimit validators (4), but it can spike to a max of 8704 in a
+   * single epoch if there haven't been exits in a while and there's a massive attester slashing at once of validators
+   * that happen to be in the same committee, which is very unlikely.
+   */
+  indicesToSlash: ValidatorIndex[];
+  /**
+   * Indices of validators that just joinned and will be eligible for the active queue.
+   * ```
+   * v.activationEligibilityEpoch === FAR_FUTURE_EPOCH && v.effectiveBalance === MAX_EFFECTIVE_BALANCE
+   * ```
+   * All validators in indicesEligibleForActivationQueue get activationEligibilityEpoch set. So it can only include
+   * validators that have just joinned the registry through a valid full deposit(s).
+   * ```
+   * max indicesEligibleForActivationQueue = SLOTS_PER_EPOCH * MAX_DEPOSITS
+   * ```
+   * For mainnet spec = 512
+   */
+  indicesEligibleForActivationQueue: ValidatorIndex[];
+  /**
+   * Indices of validators that may become active once churn and finaly allow.
+   * ```
+   * v.activationEpoch === FAR_FUTURE_EPOCH && v.activationEligibilityEpoch <= currentEpoch
+   * ```
+   * Many validators could be on indicesEligibleForActivation, but only up to churnLimit will be activated.
+   * For less than 327680 validators, churnLimit = 4 (minimum possible), so max processed is 4.
+   */
+  indicesEligibleForActivation: ValidatorIndex[];
+  /**
+   * Indices of validators that will be ejected due to low balance.
+   * ```
+   * status.active && v.exitEpoch === FAR_FUTURE_EPOCH && v.effectiveBalance <= config.EJECTION_BALANCE
+   * ```
+   * Potentially the entire validator set could be added to indicesToEject, and all validators in the array will have
+   * their validator object mutated. Exit queue churn delays exit, but the object is mutated immediately.
+   */
+  indicesToEject: ValidatorIndex[];
+
+  statuses: IAttesterStatus[];
+  balances?: number[];
+  /**
+   * Active validator indices for currentEpoch + 2.
+   * This is only used in `afterProcessEpoch` to compute epoch shuffling, it's not efficient to calculate it at that time
+   * since it requires 1 loop through validator.
+   * | epoch process fn                 | nextEpochTotalActiveBalance action |
+   * | -------------------------------- | ---------------------------------- |
+   * | beforeProcessEpoch               | calculate during the validator loop|
+   * | afterEpochProcess                | read it                            |
+   */
+  nextEpochShufflingActiveValidatorIndices: ValidatorIndex[];
+  /**
+   * Altair specific, this is total active balances for the next epoch.
+   * This is only used in `afterProcessEpoch` to compute base reward and sync participant reward.
+   * It's not efficient to calculate it at that time since it requires looping through all active validators,
+   * so we should calculate it during `processEffectiveBalancesUpdate` which gives us updated effective balance.
+   * | epoch process fn                 | nextEpochTotalActiveBalance action |
+   * | -------------------------------- | ---------------------------------- |
+   * | beforeProcessEpoch               | initialize as BigInt(0)            |
+   * | processEffectiveBalancesUpdate   | calculate during the loop          |
+   * | afterEpochProcess                | read it                            |
+   */
+  nextEpochTotalActiveBalanceByIncrement: number;
+
+  /**
+   * Track by validator index if it's active in the next epoch.
+   * Used in `processEffectiveBalanceUpdates` to save one loop over validators after epoch process.
+   */
+  isActiveNextEpoch: boolean[];
+}
