+++ packages/beacon-state-transition/src/cache/pubkeyCache.ts
