+++ packages/beacon-state-transition/src/cache/syncCommitteeCache.ts
