+++ packages/beacon-state-transition/src/util/attesterStatus.ts
