+++ packages/beacon-state-transition/src/util/unsafeUint8ArrayToTree.ts
